refactor(jira): move configuration types into jira package
